Updated Geiss scene
Played around with diffusion effect.

+++ Assets/Geiss/Unlit-Alpha-Geiss.shader
@@ -41,7 +41,7 @@ SubShader {
 
             sampler2D _MainTex;
             float4 _MainTex_ST;
-			fixed4 _Color;
+			float4 _Color;
 
             v2f vert (appdata_t v)
             {
